Extra large Christmas update:
Some of the sector fields were not saved in savegame.

Implemented support for Boom's remembering of the last enemy.

Implemented Boom's touching sectors list, use it in some places.

Fixed building of stairs if starting sector is marked as part of the staircase.

While traveling between maps in a hub should not show the intermission screen unless it's a deathmatch.

Fixed endless loop in InterpolationPoint.ScanForLoop when starting point is not part of the loop.

Added spark thing types.

Implemented HealThing line special.

Implemented support for monster patrol points and Thing_SetGoal line special.

Implemented support for using TITLEMAP in the place of the title picture.

ThingCount should not count things in somebody's inventory.

Fixed a memory leak in OpenGL driver.

Started to implement support for DECORATE scripts. Implemented support for old style simple decorations and breakables.

git-svn-id: svn://svn.code.sf.net/p/vavoom/code/trunk@2965 4be913c3-b711-0410-b476-fa31da1fb1f1